Class abstraction for core api
