+++ platform_sdk/core_api/core_metadata.py
@@ -1,0 +1,6 @@
+from core_api.core_base_entity import CoreBaseEntity
+
+class Metadata(CoreBaseEntity):
+
+    def __init__(self,url):
+        super.__init__(url,'operations')
