+++ platform_sdk/core_api/core_system_solution.py
@@ -1,0 +1,7 @@
+from core_api.core_base_entity import CoreBaseEntity
+
+class SystemSolution(CoreBaseEntity):
+
+    def __init__(self,url):
+        super.__init__(url,'system')
+
